requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
httpx==0.28.0
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.17
python-dotenv==1.0.1
discord.py==2.4.0
